@@ -6,7 +6,7 @@ if(${CMAKE_VERSION} VERSION_LESS 3.12)
 endif()
 
 if (NOT CMAKE_CXX_STANDARD)
-  set(CMAKE_CXX_STANDARD 14)
+  set(CMAKE_CXX_STANDARD 17)
 endif()
 
 if(${CMAKE_VERSION} VERSION_EQUAL "3.11.0" OR ${CMAKE_VERSION} VERSION_GREATER "3.11.0")
@@ -25,28 +25,19 @@ find_package(ament_cmake REQUIRED)
 find_package(cv_bridge REQUIRED)
 find_package(gps_msgs REQUIRED)
 find_package(mapviz REQUIRED)
+find_package(OpenGL REQUIRED)
 find_package(pluginlib REQUIRED)
+find_package(Qt5Core REQUIRED)
+find_package(Qt5Gui REQUIRED)
+find_package(Qt5OpenGL REQUIRED)
+find_package(Qt5Widgets REQUIRED)
 find_package(rclcpp  REQUIRED)
 find_package(tf2 REQUIRED)
 find_package(tf2_ros REQUIRED)
 find_package(swri_transform_util  REQUIRED)
 find_package(swri_math_util REQUIRED)
 
-### Boost ###
-find_package(Boost REQUIRED filesystem thread)
-
-### OpenGL ###
-find_package(OpenGL REQUIRED)
-
 ### OpenCV ###
-# multires_image doesn't directly use OpenCV, but it uses swri_transform_util, and
-# we need to include OpenCV to make one of its headers compile.
-find_package(OpenCV COMPONENTS core imgproc REQUIRED)
-
-find_package(Qt5Core REQUIRED)
-find_package(Qt5Gui REQUIRED)
-find_package(Qt5OpenGL REQUIRED)
-find_package(Qt5Widgets REQUIRED)
 add_definitions(-DWFlags=WindowFlags)
 
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
@@ -65,23 +56,23 @@ add_library(${PROJECT_NAME}
 
 target_include_directories(${PROJECT_NAME}
   PUBLIC
-    include
-    ${swri_math_util_INCLUDE_DIRS}
-    ${swri_transform_util_INCLUDE_DIRS})
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:include/${PROJECT_NAME}>
+)
 
 target_link_libraries(${PROJECT_NAME} PUBLIC
-  ${Boost_LIBRARIES}
   cv_bridge::cv_bridge
   OpenGL::GL
   ${gps_msgs_TARGETS}
+  ${mapviz_TARGETS}
   pluginlib::pluginlib
   Qt5::Core
   Qt5::Gui
   Qt5::OpenGL
   Qt5::Widgets
   rclcpp::rclcpp
-  ${swri_math_util_LIBRARIES}
-  ${swri_transform_util_LIBRARIES}
+  ${swri_math_util_TARGETS}
+  ${swri_transform_util_TARGETS}
   tf2::tf2
   tf2_ros::tf2_ros
 )
@@ -114,8 +105,6 @@ qt5_wrap_cpp(VIEW_NODE_SRC_FILES include/${PROJECT_NAME}/multires_view_node.h)
 add_executable(multires_view_node ${VIEW_NODE_SRC_FILES})
 target_link_libraries(multires_view_node PUBLIC
   multires_widget
-  Boost::filesystem
-  Boost::thread
   rclcpp::rclcpp
 )
 
@@ -129,22 +118,30 @@ qt5_wrap_cpp(MAPVIZ_SRC_FILES include/${PROJECT_NAME}/${PROJECT_NAME}_plugin.h)
 add_library(${PROJECT_NAME}_plugin SHARED
   ${MAPVIZ_SRC_FILES}
 )
-target_compile_definitions(${PROJECT_NAME}_plugin PUBLIC "PLUGINLIB__DISABLE_BOOST_FUNCTIONS")
-target_link_libraries(${PROJECT_NAME}_plugin
+
+target_link_libraries(${PROJECT_NAME}_plugin PUBLIC
   ${PROJECT_NAME}
+  multires_widget
+  ${mapviz_TARGETS}
 )
 
 ### Install ${PROJECT_NAME} plugin ###
-install(DIRECTORY include/${PROJECT_NAME}/
+install(DIRECTORY include/
   DESTINATION include/${PROJECT_NAME}
   FILES_MATCHING PATTERN "*.h"
 )
 
+# Install executables/nodes
+install(TARGETS multires_view_node
+  DESTINATION lib/${PROJECT_NAME}
+)
+
+# Install and export libraries
 install(TARGETS ${PROJECT_NAME}
-  multires_view_node
   multires_widget
   ${PROJECT_NAME}_plugin
-  RUNTIME DESTINATION lib/${PROJECT_NAME}
+  EXPORT export_${PROJECT_NAME}
+  RUNTIME DESTINATION bin
   LIBRARY DESTINATION lib
   ARCHIVE DESTINATION lib
 )
@@ -160,7 +157,17 @@ install(DIRECTORY launch
 
 pluginlib_export_plugin_description_file(mapviz mapviz_plugins.xml)
 
-ament_export_dependencies(${RUNTIME_DEPS} Qt)
-ament_export_include_directories(include)
-ament_export_libraries(${PROJECT_NAME})
+ament_export_targets(export_${PROJECT_NAME})
+ament_export_dependencies(${RUNTIME_DEPS} 
+  Qt5Core
+  Qt5Gui
+  Qt5OpenGL
+  Qt5Widgets
+)
+ament_export_include_directories(
+  include/${PROJECT_NAME}
+)
+ament_export_libraries(
+  ${PROJECT_NAME}
+)
 ament_package()
