@@ -6,7 +6,7 @@ if(${CMAKE_VERSION} VERSION_LESS 3.12)
 endif()
 
 if (NOT CMAKE_CXX_STANDARD)
-  set(CMAKE_CXX_STANDARD 14)
+  set(CMAKE_CXX_STANDARD 17)
 endif()
 
 if(${CMAKE_VERSION} VERSION_EQUAL "3.11.0" OR ${CMAKE_VERSION} VERSION_GREATER "3.11.0")
@@ -37,8 +37,6 @@ find_package(Qt5Gui REQUIRED)
 find_package(Qt5Network REQUIRED)
 find_package(Qt5OpenGL REQUIRED)
 find_package(Qt5Widgets REQUIRED)
-# Still useful?
-# add_definitions(-DWFlags=WindowFlags)
 
 ### OpenGL ###
 find_package(OpenGL REQUIRED)
@@ -54,9 +52,6 @@ pkg_check_modules(YamlCpp yaml-cpp)
 
 ### Use PkgConfig to find jsoncpp ###
 pkg_check_modules(JSONCPP REQUIRED jsoncpp)
-
-# Fix conflict between Boost signals used by tf and QT signals used by mapviz
-add_definitions(-DQT_NO_KEYWORDS)
 
 # Need to include the current dir to include the results of building Qt UI files
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
@@ -79,24 +74,24 @@ qt5_wrap_cpp(TILE_SRC_FILES ${QT_HEADERS})
 add_library(${PROJECT_NAME} ${TILE_SRC_FILES})
 target_include_directories(${PROJECT_NAME}
   PUBLIC
-    include
-    GLEW::GLEW
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:include/${PROJECT_NAME}>
     ${JSONCPP_INCLUDE_DIRS}
-    ${mapviz_INCLUDE_DIRS}
-    ${swri_transform_util_INCLUDE_DIRS}
     ${YamlCpp_INCLUDE_DIRS})
-
+    
 target_link_libraries(${PROJECT_NAME} PUBLIC
   ${GLU_LIBRARY}
+  GLEW::GLEW
   ${JSONCPP_LIBRARIES}
+  ${mapviz_TARGETS}
   Qt5::Network
   Qt5::Core
   Qt5::Gui
   Qt5::OpenGL
   Qt5::Widgets
-  ${mapviz_LIBRARIES}
   pluginlib::pluginlib
   rclcpp::rclcpp
+  ${swri_transform_util_TARGETS}
   tf2::tf2
   tf2_ros::tf2_ros
   ${YamlCpp_LIBRARIES})
@@ -110,22 +105,23 @@ qt5_wrap_ui(PLUGIN_SRC_FILES ${PLUGIN_UI_FILES})
 qt5_wrap_cpp(PLUGIN_SRC_FILES include/${PROJECT_NAME}/${PROJECT_NAME}_plugin.h)
 
 add_library(${PROJECT_NAME}_plugin SHARED ${PLUGIN_SRC_FILES})
-target_compile_definitions(${PROJECT_NAME}_plugin PUBLIC "PLUGINLIB__DISABLE_BOOST_FUNCTIONS")
 target_link_libraries(${PROJECT_NAME}_plugin ${PROJECT_NAME})
 
-install(DIRECTORY include/${PROJECT_NAME}/
+install(DIRECTORY include
   DESTINATION include/${PROJECT_NAME}
   FILES_MATCHING PATTERN "*.h"
 )
 
 install(TARGETS ${PROJECT_NAME} ${PROJECT_NAME}_plugin
-  RUNTIME DESTINATION lib/${PROJECT_NAME}
+  EXPORT export_${PROJECT_NAME}
+  RUNTIME DESTINATION bin
   LIBRARY DESTINATION lib
   ARCHIVE DESTINATION lib
 )
 
 pluginlib_export_plugin_description_file(mapviz mapviz_plugins.xml)
 
+ament_export_targets(export_${PROJECT_NAME})
 ament_export_dependencies(${RUNTIME_DEPS}
   Qt5::Network
   Qt5::Core
@@ -133,6 +129,4 @@ ament_export_dependencies(${RUNTIME_DEPS}
   Qt5::OpenGL
   Qt5::Widgets
 )
-ament_export_include_directories(include)
-ament_export_libraries(${PROJECT_NAME})
 ament_package()
