@@ -32,9 +32,7 @@
 
 // C++ standard libraries
 #include <string>
-
-// Boost libraries
-#include <boost/filesystem.hpp>
+#include <filesystem>
 
 #include <mapviz/mapviz_plugin.h>
 
@@ -104,9 +102,9 @@ namespace mapviz_plugins
 
     void GetCenterPoint(double x, double y);
 
-    boost::filesystem::path MakePathRelative(
-      boost::filesystem::path path,
-      boost::filesystem::path base);
+    std::filesystem::path MakePathRelative(
+      const std::filesystem::path& path,
+      const std::filesystem::path& base);
   };
 }
 
