@@ -36,6 +36,7 @@
  */
 
 #include <string>
+#include <filesystem>
 
 // QT libraries
 #include <QString>
@@ -53,7 +54,7 @@ namespace multires_image
   , argc_(argc)
   , argv_(argv)
   , node_(nullptr)
-  , thread_(nullptr)
+  , thread_{}
   , initialized_(false)
   , tile_set_(nullptr)
   {
@@ -65,7 +66,7 @@ namespace multires_image
   {
     if (!thread_)
     {
-      thread_ = new boost::thread(&MultiresViewNode::SpinLoop, this);
+      thread_ = new std::thread(&MultiresViewNode::SpinLoop, this);
     }
   }
 
