@@ -358,8 +358,32 @@ private:
 };
 typedef std::shared_ptr<MapvizPlugin> MapvizPluginPtr;
 
-// Implementation
+// Dealing with YAML frequently requires trimming whitespace from strings
+inline std::string TrimString(const std::string& str)
+{
+  auto begin = str.begin();
+  auto end = str.end();
 
+  // Trim leading whitespace
+  while (begin != end && std::isspace(*begin))
+  {
+    ++begin;
+  }
+
+  // Trim trailing whitespace
+  if (begin != end)
+  {
+    do
+    {
+      --end;
+    } while (std::isspace(*end));
+    ++end;
+  }
+
+  return std::string(begin, end);
+}
+
+// Implementation
 inline void MapvizPlugin::PrintErrorHelper(QLabel *status_label, const std::string &message,
                                             double throttle)
 {
