@@ -30,9 +30,6 @@
 #include <mapviz_plugins/textured_marker_plugin.h>
 #include <mapviz_plugins/topic_select.h>
 
-// Boost libraries
-#include <boost/algorithm/string.hpp>
-
 // QT libraries
 #include <QDialog>
 #include <QGLWidget>
@@ -535,7 +532,8 @@ void TexturedMarkerPlugin::LoadConfig(const YAML::Node & node, const std::string
   LoadQosConfig(node, qos_);
   if (node["topic"]) {
     std::string topic = node["topic"].as<std::string>();
-    ui_.topic->setText(boost::trim_copy(topic).c_str());
+    topic.erase(std::remove_if(topic.begin(), topic.end(), ::isspace), topic.end());
+    ui_.topic->setText(topic.c_str());
   }
 
   TopicEdited();
@@ -543,8 +541,9 @@ void TexturedMarkerPlugin::LoadConfig(const YAML::Node & node, const std::string
 
 void TexturedMarkerPlugin::SaveConfig(YAML::Emitter & emitter, const std::string & path)
 {
-  emitter << YAML::Key << "topic" << YAML::Value <<
-    boost::trim_copy(ui_.topic->text().toStdString());
+  std::string topic = ui_.topic->text().toStdString();
+    topic.erase(std::remove_if(topic.begin(), topic.end(), ::isspace), topic.end());
+  emitter << YAML::Key << "topic" << YAML::Value << topic;
 
   SaveQosConfig(emitter, qos_);
 }
