@@ -1,6 +1,6 @@
 // *****************************************************************************
 //
-// Copyright (c) 2015, Southwest Research Institute® (SwRI®)
+// Copyright (c) 2015-2025, Southwest Research Institute® (SwRI®)
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -33,7 +33,7 @@
 
 #include "tile_source.h"
 
-#include <boost/functional/hash.hpp>
+#include <string>
 
 namespace tile_map
 {
@@ -79,7 +79,7 @@ namespace tile_map
     static const QString WMTS_TYPE;
 
   private:
-    boost::hash<std::string> hash_;
+    std::hash<std::string> hash_;
   };
 }
 
