@@ -1,6 +1,6 @@
 // *****************************************************************************
 //
-// Copyright (c) 2015, Southwest Research Institute® (SwRI®)
+// Copyright (c) 2015-2025, Southwest Research Institute® (SwRI®)
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -29,8 +29,6 @@
 // *****************************************************************************
 
 #include <tile_map/wmts_source.h>
-
-#include <boost/functional/hash.hpp>
 
 namespace tile_map
 {
